#9640 Refactor: Move util::str::is_token to script::dom::bindings::str
