+++ components/script/dom/bindings/str.rs
@@ -10,7 +10,6 @@ use std::hash::{Hash, Hasher};
 use std::ops;
 use std::str;
 use std::str::FromStr;
-use util::str::is_token;
 
 /// Encapsulates the IDL `ByteString` type.
 #[derive(JSTraceable, Clone, Eq, PartialEq, HeapSizeOf)]
@@ -142,3 +141,38 @@ impl ops::Deref for ByteString {
 /// A string that is constructed from a UCS-2 buffer by replacing invalid code
 /// points with the replacement character.
 pub struct USVString(pub String);
+
+
+/// Returns whether `s` is a `token`, as defined by
+/// [RFC 2616](http://tools.ietf.org/html/rfc2616#page-17).
+pub fn is_token(s: &[u8]) -> bool {
+    if s.is_empty() {
+        return false; // A token must be at least a single character
+    }
+    s.iter().all(|&x| {
+        // http://tools.ietf.org/html/rfc2616#section-2.2
+        match x {
+            0...31 | 127 => false, // CTLs
+            40 |
+            41 |
+            60 |
+            62 |
+            64 |
+            44 |
+            59 |
+            58 |
+            92 |
+            34 |
+            47 |
+            91 |
+            93 |
+            63 |
+            61 |
+            123 |
+            125 |
+            32 => false, // separators
+            x if x > 127 => false, // non-CHARs
+            _ => true,
+        }
+    })
+}
