+++ components/script/dom/websocket.rs
@@ -16,7 +16,7 @@ use dom::bindings::inheritance::Castable;
 use dom::bindings::js::Root;
 use dom::bindings::refcounted::Trusted;
 use dom::bindings::reflector::{Reflectable, reflect_dom_object};
-use dom::bindings::str::USVString;
+use dom::bindings::str::{USVString, is_token};
 use dom::bindings::trace::JSTraceable;
 use dom::blob::Blob;
 use dom::closeevent::CloseEvent;
@@ -41,7 +41,7 @@ use std::borrow::ToOwned;
 use std::cell::Cell;
 use std::ptr;
 use std::thread;
-use util::str::{DOMString, is_token};
+use util::str::DOMString;
 use websocket::client::request::Url;
 use websocket::header::{Headers, WebSocketProtocol};
 use websocket::ws::util::url::parse_url;
